+++ src/com/maddyhome/idea/vim/group/KeyGroup.java
@@ -35,6 +35,7 @@ import com.maddyhome.idea.vim.command.Argument;
 import com.maddyhome.idea.vim.command.Command;
 import com.maddyhome.idea.vim.command.MappingMode;
 import com.maddyhome.idea.vim.ex.ExOutputModel;
+import com.maddyhome.idea.vim.extension.VimExtensionHandler;
 import com.maddyhome.idea.vim.helper.StringHelper;
 import com.maddyhome.idea.vim.key.*;
 import com.maddyhome.idea.vim.key.Shortcut;
@@ -64,6 +65,7 @@ public class KeyGroup {
   @NotNull private final Set<KeyStroke> requiredShortcutKeys = new HashSet<KeyStroke>();
   @NotNull private final HashMap<MappingMode, RootNode> keyRoots = new HashMap<MappingMode, RootNode>();
   @NotNull private final Map<MappingMode, KeyMapping> keyMappings = new HashMap<MappingMode, KeyMapping>();
+  @Nullable private OperatorFunction operatorFunction = null;
 
   public void registerRequiredShortcutKeys(@NotNull Editor editor) {
     final Set<KeyStroke> requiredKeys = VimPlugin.getKey().getRequiredShortcutKeys();
@@ -85,7 +87,18 @@ public class KeyGroup {
       builder.append(" ");
       builder.append(row.isRecursive() ? " " : "*");
       builder.append(" ");
-      builder.append(toKeyNotation(row.getToKeys()));
+      final List<KeyStroke> toKeys = row.getToKeys();
+      final VimExtensionHandler extensionHandler = row.getExtensionHandler();
+      if (toKeys != null) {
+        builder.append(toKeyNotation(toKeys));
+      }
+      else if (extensionHandler != null) {
+        builder.append("call ");
+        builder.append(extensionHandler.getClass().getCanonicalName());
+      }
+      else {
+        builder.append("<Unknown>");
+      }
       builder.append("\n");
     }
     ExOutputModel.getInstance(editor).output(builder.toString());
@@ -93,10 +106,11 @@ public class KeyGroup {
   }
 
   public void putKeyMapping(@NotNull Set<MappingMode> modes, @NotNull List<KeyStroke> fromKeys,
-                            @NotNull List<KeyStroke> toKeys, boolean recursive) {
+                            @Nullable List<KeyStroke> toKeys, @Nullable VimExtensionHandler extensionHandler,
+                            boolean recursive) {
     for (MappingMode mode : modes) {
       final KeyMapping mapping = getKeyMapping(mode);
-      mapping.put(EnumSet.of(mode), fromKeys, toKeys, recursive);
+      mapping.put(EnumSet.of(mode), fromKeys, toKeys, extensionHandler, recursive);
     }
     final int oldSize = requiredShortcutKeys.size();
     for (KeyStroke key : fromKeys) {
@@ -110,6 +124,15 @@ public class KeyGroup {
         registerRequiredShortcutKeys(editor);
       }
     }
+  }
+
+  @Nullable
+  public OperatorFunction getOperatorFunction() {
+    return operatorFunction;
+  }
+
+  public void setOperatorFunction(@NotNull OperatorFunction function) {
+    operatorFunction = function;
   }
 
   public void saveData(@NotNull Element element) {
@@ -411,7 +434,7 @@ public class KeyGroup {
         final MappingInfo mappingInfo = mapping.get(fromKeys);
         if (mappingInfo != null) {
           rows.add(new MappingInfo(mappingModes, mappingInfo.getFromKeys(), mappingInfo.getToKeys(),
-                                   mappingInfo.isRecursive()));
+                                   mappingInfo.getExtensionHandler(), mappingInfo.isRecursive()));
         }
       }
     }